# Talk-Leee/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Talk-Lee API",
    "version": "0.1.0",
    "description": "Generated from frontend client contracts (schemas + endpoint registry)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Connector": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "config": {
            "type": "object",
            "additionalProperties": {}
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "config",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
      "ConnectorListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "config": {
                  "type": "object",
                  "additionalProperties": {}
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "type",
                "config",
                "createdAt"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "detail": {
            "type": "string",
            "examples": [
              "Unauthorized"
            ]
          }
        },
        "required": [
          "detail"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/connectors": {
      "get": {
        "tags": [
          "Connectors"
        ],
        "summary": "List connectors",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Connectors"
        ],
        "summary": "Create connector",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string",
                    "examples": [
                      "Google Calendar"
                    ]
                  },
                  "type": {
                    "type": "string",
                    "examples": [
                      "calendar"
                    ]
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": true,
                    "examples": [
                      {
                        "clientId": "..."
                      }
                    ]
                  }
                },
                "required": [
                  "name",
                  "type",
                  "config"
                ]
              },
              "examples": {
                "create": {
                  "value": {
                    "name": "Google Calendar",
                    "type": "calendar",
                    "config": {
                      "clientId": "..."
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connector"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}